LJT-101 Seperate unit and integration tests (#109)
* LJT-101 Seperate unit and integration tests

* Move unit tests to corresponding directory

* Add missing imports

* Fix issue where tests repeat twice

* Change from buildDir to layout.buildDirectory
