+++ src/test/java/com/coherentsolutions/pot/insuranceservice/unit/auth0/Auth0UserDtoTest.java
@@ -1,10 +1,11 @@
-package com.coherentsolutions.pot.insuranceservice.dto.auth0;
+package com.coherentsolutions.pot.insuranceservice.unit.auth0;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import com.coherentsolutions.pot.insuranceservice.dto.auth0.Auth0UserDto;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
