+++ src/test/java/com/coherentsolutions/pot/insuranceservice/unit/config/Auth0PropertiesTest.java
@@ -1,8 +1,9 @@
-package com.coherentsolutions.pot.insuranceservice.config;
+package com.coherentsolutions.pot.insuranceservice.unit.config;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
+import com.coherentsolutions.pot.insuranceservice.config.Auth0Properties;
 import org.junit.jupiter.api.Test;
 
 /**
