+++ src/test/java/com/coherentsolutions/pot/insuranceservice/unit/service/Auth0UserManagementServiceTest.java
@@ -1,4 +1,4 @@
-package com.coherentsolutions.pot.insuranceservice.service;
+package com.coherentsolutions.pot.insuranceservice.unit.service;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -20,6 +20,7 @@ import com.auth0.net.Request;
 import com.auth0.net.Response;
 import com.coherentsolutions.pot.insuranceservice.dto.auth0.Auth0UserDto;
 import com.coherentsolutions.pot.insuranceservice.mapper.Auth0UserMapper;
+import com.coherentsolutions.pot.insuranceservice.service.Auth0UserManagementService;
 import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
